+++ .idea/misc.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="CMakeWorkspace" PROJECT_DIR="$PROJECT_DIR$/IZ_2">
+    <contentRoot DIR="$PROJECT_DIR$" />
+  </component>
+</project>
