+++ IZ_2/CMakeLists.txt
@@ -12,15 +12,15 @@ add_subdirectory(lib/utils)
 add_subdirectory(lib/matrix)
 add_subdirectory(lib/errors)
 add_subdirectory(lib/matrix_reverse)
+add_subdirectory(lib/matrix_reverse_parallel_copy_interface)
 add_subdirectory(lib/matrix_reverse_parallel)
 add_subdirectory(lib/stress_tests)
-add_subdirectory(lib/matrix_reverse_parallel_for_main)
 
-set(CHOOSE_MODE 0) #1 - parallel 0 - linear
-if(CHOOSE_MODE STREQUAL 1)
-    add_executable(main src/main.c)
-    target_link_libraries(main matrix matrix_reverse_parallel_shared_for_main)
-else(CHOOSE_MODE STREQUAL 0)
+set(CHOOSE_MODE 0) #0 - linear else - parallel
+if(CHOOSE_MODE STREQUAL 0)
     add_executable(main src/main.c)
     target_link_libraries(main matrix matrix_reverse)
+else()
+    add_executable(main src/main.c)
+    target_link_libraries(main matrix matrix_reverse_parallel_copy_interface)
 endif()
