+++ IZ_2/lib/errors/include/errors.h
@@ -1,8 +1,8 @@
 #ifndef ERRORS_H
 #define ERRORS_H
 
-typedef enum {
-    ERROR_OK = 0,
+typedef enum{
+    OK = 0,
     ERROR_NULL_PTR_REFERENCE,
     ERROR_MEMORY_ALLOCATION_ERROR,
     ERROR_OUT_RANGE_SET,
