+++ IZ_2/lib/matrix/tests/src/main_test.cpp
@@ -5,7 +5,7 @@ extern "C"{
 #include "matrix_reverse.h"
 }
 
-TEST(init_matrix, init_matrix) {
+TEST(init_matrix, init_matrix){
     matrix_t matrix;
     init_matrix(&matrix);
     EXPECT_EQ(matrix.array, (void *)NULL);
@@ -13,7 +13,7 @@ TEST(init_matrix, init_matrix) {
     EXPECT_EQ(matrix.size_y, 0);
 }
 
-TEST(init_matrix, init_matrix_null) {
+TEST(init_matrix, init_matrix_null){
     matrix_t *matrix = NULL;
     init_matrix(matrix);
 }
@@ -24,7 +24,7 @@ TEST(set_elem, set_elem){
     new_matrix(&matrix, 10, 10);
     EXPECT_EQ(matrix.array[0], 0);
     matrix_error_t error = set_elem(&matrix, 0, 0, 1);
-    EXPECT_EQ(error, ERROR_OK);
+    EXPECT_EQ(error, OK);
     EXPECT_EQ(matrix.array[0], 1);
     free_matrix(&matrix);
 }
@@ -48,15 +48,15 @@ TEST(set_matrix_rand_elem_1_9, set_matrix_rand_elem_1_9){
     init_matrix(&matrix);
     matrix_error_t error;
     error = new_matrix(&matrix, 1000, 1000);
-    EXPECT_EQ(error, ERROR_OK);
+    EXPECT_EQ(error, OK);
     error = set_matrix_rand_elem_1_9(&matrix);
-    EXPECT_EQ(error, ERROR_OK);
-    for (size_t i = 0; i < matrix.size_x * matrix.size_y; ++i) {
+    EXPECT_EQ(error, OK);
+    for (size_t i = 0; i < matrix.size_x * matrix.size_y; ++i){
         EXPECT_GT(matrix.array[i], 0);
         EXPECT_LT(matrix.array[i], 10);
     }
     error = free_matrix(&matrix);
-    EXPECT_EQ(error, ERROR_OK);
+    EXPECT_EQ(error, OK);
 }
 
 TEST(set_matrix_rand_elem_1_9, set_matrix_rand_elem_1_9_null){
@@ -66,7 +66,7 @@ TEST(set_matrix_rand_elem_1_9, set_matrix_rand_elem_1_9_null){
 }
 
 
-int main(int argc, char **argv) {
+int main(int argc, char **argv){
     ::testing::InitGoogleTest(&argc, argv);
     return RUN_ALL_TESTS();
 }
