+++ IZ_2/lib/matrix_reverse_parallel/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library(matrix_reverse_parallel_shared SHARED src/matrix_reverse_parallel_for_main.c)
+add_library(matrix_reverse_parallel_shared SHARED src/matrix_reverse_parallel.c)
 target_include_directories(matrix_reverse_parallel_shared PUBLIC include)
 target_link_libraries(matrix_reverse_parallel_shared PUBLIC errors)
 target_link_libraries(matrix_reverse_parallel_shared PUBLIC matrix)
