+++ IZ_2/lib/matrix_reverse_parallel_copy_interface/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_library(matrix_reverse_parallel_copy_interface SHARED src/matrix_reverse_parallel_copy_interface.c)
+target_include_directories(matrix_reverse_parallel_copy_interface PUBLIC include)
+target_link_libraries(matrix_reverse_parallel_copy_interface PUBLIC matrix_reverse_parallel_shared)
+target_link_libraries(matrix_reverse_parallel_copy_interface PUBLIC errors)
+target_link_libraries(matrix_reverse_parallel_copy_interface PUBLIC matrix)
+target_link_libraries(matrix_reverse_parallel_copy_interface PUBLIC utils)
