+++ IZ_2/lib/stress_tests/src/main_test.cpp
@@ -8,7 +8,7 @@ extern "C"{
 }
 
 TEST(stress_tests, stress_tests){
-    for(int k = 0; k < 4; k++) {
+    for(int k = 0; k < 4; k++){
         matrix_t matrix1, matrix2;
         init_matrix(&matrix1);
         init_matrix(&matrix2);
@@ -16,17 +16,17 @@ TEST(stress_tests, stress_tests){
         size_t size_y = rand() % 10000;
         matrix_error_t error;
         error = new_matrix_mmap(&matrix1, size_x, size_y);
-        EXPECT_EQ(error, ERROR_OK);
+        EXPECT_EQ(error, OK);
         error = new_matrix(&matrix2, size_x, size_y);
-        EXPECT_EQ(error, ERROR_OK);
+        EXPECT_EQ(error, OK);
         int val = 0;
-        for (size_t i = 0; i < size_y; ++i) {
-            for (size_t j = 0; j < size_x; j++) {
+        for (size_t i = 0; i < size_y; ++i){
+            for (size_t j = 0; j < size_x; j++){
                 val = int(rand() % 10000);
                 error = set_elem(&matrix1, j, i, val);
-                EXPECT_EQ(error, ERROR_OK);
+                EXPECT_EQ(error, OK);
                 error = set_elem(&matrix2, j, i, val);
-                EXPECT_EQ(error, ERROR_OK);
+                EXPECT_EQ(error, OK);
             }
         }
 
@@ -37,14 +37,14 @@ TEST(stress_tests, stress_tests){
             EXPECT_EQ(matrix1.array[i], matrix2.array[i]);
 
         error = free_matrix_mmap(&matrix1);
-        EXPECT_EQ(error, ERROR_OK);
+        EXPECT_EQ(error, OK);
         error = free_matrix(&matrix2);
-        EXPECT_EQ(error, ERROR_OK);
+        EXPECT_EQ(error, OK);
     }
 }
 
 
-int main(int argc, char **argv) {
+int main(int argc, char **argv){
     ::testing::InitGoogleTest(&argc, argv);
     return RUN_ALL_TESTS();
 }
